bot/pyproject.toml
[tool.poetry]
name = "bot"
version = "0.1.0"
description = ""
authors = ["Dilemma25 <valtert90@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
aiogram = "^3.20.0"  
python-dotenv = "^1.1.0"
redis = "^6.2.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"